project/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pdfplumber>=0.7.6
python-docx>=0.8.11
requests>=2.28.0
pydantic>=1.10.0
pytest>=7.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
